Replacing regexp with vilter_var(). Add support for IPv6
